project directory setup
