+++ App/App.js
@@ -1,31 +1,7 @@
-import * as React from 'react';
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View, Button } from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Page from './pages/index.js';
-
-const Stack = createNativeStackNavigator();
-
-function HomeScreen({ navigation }) {
-    return (
-        <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
-            <Text>Home Screen</Text>
-            <Button
-                title="Go to Details"
-                onPress={() => navigation.navigate('Page')}
-            />
-        </View>
-    );
-}
+import Navigation from './pages/navigation';
 
 export default function App() {
     return (
-        <NavigationContainer>
-            <Stack.Navigator>
-                <Stack.Screen name="Home" component={HomeScreen} />
-                <Stack.Screen name="Page" component={Page} />
-            </Stack.Navigator>
-        </NavigationContainer>
+        <Navigation />
     );
 }
