+++ App/package.json
@@ -10,13 +10,14 @@
   },
   "dependencies": {
     "@react-navigation/native": "^6.1.14",
-    "@react-navigation/native-stack": "^6.9.22",
+    "@react-navigation/native-stack": "^6.9.25",
     "axios": "^1.6.7",
-    "expo": "~50.0.11",
+    "expo": "~50.0.13",
     "expo-router": "^3.4.8",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
     "react-native": "0.73.4",
+    "react-native-dotenv": "^3.4.11",
     "react-native-safe-area-context": "^4.8.2",
     "react-native-screens": "~3.29.0",
     "react-navigation": "^5.0.0"
